Fix logging crash issue and Android 11 storage permission support for debug builds

+++ core-lib/src/main/java/org/idpass/smartscanner/lib/platform/utils/LoggerUtils.kt
@@ -22,16 +22,22 @@ import java.io.File
 
 object LoggerUtils {
     fun writeLogToFile(identifier : String = "default") {
-        FileUtils.createSmartScannerDirs()
-        val fileName = "logcat-$identifier.txt"
-        val file = File(FileUtils.directory, fileName)
-        if (!file.exists()) {
-            file.createNewFile()
-        } else {
-            file.delete()
+        try {
+            val directory = FileUtils.directory
+            val fileName = "logcat-$identifier.txt"
+            if (File(directory).exists()) {
+                val file = File(directory, fileName)
+                if (!file.exists()) {
+                    file.createNewFile()
+                } else {
+                    file.delete()
+                }
+                //"logcat -f *:S NFCScannerActivity:D SmartScannerActivity:D"
+                val command = "logcat -f " + file.absolutePath
+                Runtime.getRuntime().exec(command)
+            }
+        } catch (e : Exception) {
+            e.printStackTrace()
         }
-        //"logcat -f *:S NFCScannerActivity:D SmartScannerActivity:D"
-        val command = "logcat -f " + file.absolutePath
-        Runtime.getRuntime().exec(command)
     }
 }
